+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Add missing form data into Blueprints class
     * Allow dynamic fields in Blueprints (property@)
     * Add merging strategy support for Blueprints
+    * Implement reordering fields in Blueprints
     * Remove dependency on ircmaxell/password-compat
 2. [](#improved)
     * UniformResourceLocator::addPath(): Add option to add path after existing one (falls back to be last if path is not found)